Moved Facebook Canvas into separate submodule. Did all TODOs.
